requirements.txt
swig==4.2.1
streamlit==1.32.0
pandas==2.1.4
numpy==1.26.2
scikit-learn==1.3.2
matplotlib==3.8.2
seaborn==0.13.0
plotly==5.18.0
torch==2.2.0
transformers==4.41.0
sentence-transformers==4.1.0
langchain==0.0.348
chromadb==0.4.15
pypdf==3.17.4
google-cloud-aiplatform
faiss-cpu==1.10.0
google-generativeai==0.3.2